Removed FileHandler class
